Extract some methods to a module
* The `BaseForm` class was growing with auxiliary methods related to
the attributes. It makes sense to extract these to a separate module.

* Added a couple of new auxiliary methods to this module:
`attribute_names` and `selected_options`. Will be needed for another
refactor.

* Also, introduced a type check in the `.build` method to avoid silly
mistakes difficult to track.
